feat: further refactor
