+++ nobodywho/core/src/chat/api/mod.rs
@@ -1,3 +1,5 @@
 pub mod handle;
+pub mod message;
 
 pub use handle::{ChatBuilder, ChatConfig, ChatHandle, ChatHandleAsync};
+pub use message::{Message, Role};
