+++ nobodywho/core/src/chat/worker/context_management.rs
