+++ nobodywho/core/src/chat/worker/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod context;
+pub(crate) mod context_management;
 pub(crate) mod generation;
 
 use crate::errors::{
